simplify things to a single script, fix twitter and add timestamps
